crm114-20080317-BlameJoeLangeway_VT-Ger-194X :
    * Major fixes to the Win32 syscall command; now

        exectest.crm

      does actually run on Win32. Please note that UNIX commands still won't work allow
      of a sudden on a Win32 box - unless you provide the proper binaries and environment -
      but at least vanilla syscall, syscall <keep> and even syscall <async> now work async
      expected: no crash or indefinite lockup anymore!

      Note that this still applies:
        "Sorry, syscall to a label isn't implemented in this version "

    * Adapted error reporting to match Bill's code. See the

        bill_style_errormessage

      constant in crm_[util_]errorhandlers.c - maybe we'll make this run-time configurable
      one day, instead of compile-time like it is today...

      Bill's format puts the code info at the end, instead of at the start, and looks
      like this:

         (runtime system location: crm_expr_file_io.c(151) in routine: crm_expr_input)

    * Win32 megatest now tests OK for these classifiers (run

        ( megatest.bat 2 2>&1 ) > megatest_win32.log

      in a DOS box to verify):

          ****  Default (SBPH Markovian) classifier
          **** OSB Markovian classifier
          **** OSB Markov Unique classifier
          **** OSB Markov Chisquared Unique classifier
          **** OSBF Local Confidence (Fidelis) classifier
          **** OSB Winnow classifier
          **** Unigram Bayesian classifier
          **** unigram Winnow classifier

          **** OSB Hyperspace classifier                  <-- matches better when you define
                    #define USE_FIXED_UNIQUE_MODE     0  /* zero! trurn OFF to match Bill's! */
                WARNING: there's a VT bug in the 20080324 wget copy of CRM114, which reflects
                         on the knowngood: it is NOT good in that regard. The GerH builds
                         includes the fixes, but will differ from knowngood until the base has
                         been fixed too.

          **** OSB three-letter Hyperspace classifier     <-- as above
          **** Unigram Hyperspace classifier              <-- as above
          **** String Hyperspace classifier               <-- as above
          **** String Unigram Hyperspace classifier       <-- as above
          **** Vector 3-word-bag Hyperspace classifier    <-- as above
          **** Bit-Entropy classifier        <-- entropy numbers differ 0.00001 percent at most.
                                                 Which might be considered excellent.
          **** Fast Substring Compression Match Classifier
          **** Bytewise Correlation classifier            <--- turned out A-OK, when I finally discovered that fixing typo's etc
                                                               in documents does NOT help megatest.sh to produce identical
                                                               results. I'm a dunce.

      dubious:

          **** Bit-Entropy Toroid classifier  <-- 10% difference in pR and slight difference in jump counts. Cause yet unknown.

          **** Support Vector Machine (SVM) unigram classifier  <-- passes the test with 0.2% difference in pR tops,
                                                                    but yaks about some files in certain tests,
                                                                    due to tightened error checking in GerH.
                                                                    A 'hm'.
          **** Support Vector Machine (SVM) classifier          <-- as above. Another 'hm'.
          **** String Kernel SVM (SKS) classifier
          **** String Kernel SVM (SKS) Unique classifier        (maybe a bug in megatest_ng.sh ??? )

      dangerous:

          **** Neural Network Classifier     <-- crashes; feature counts which differ wildly between 64-bit UNIX,
                                                 knowngood and Win32.

          **** Clump / Pmulc Test            <-- if you like wierd crashes and code which shouldn't have worked in the first
                                                 place...

      Meanwhile

              echo === TEST 014 ===
              .\crm114 matchtest.crm  < matchtest_mt_ng_1.input

      still needs some medical attention IMHO, as CRM114 on Win32 is much more
      'picky' about whitespace than its UNIX brethren.  Why? Don't know yet.
      But it shown in line leading and trailing whitespace sometimes.

      Probably due to the clunky whitespace skip loops in the CRM114 code.

    * UNIX / GNU make now comes with a working 'test' target, which actually uses the new
      megatest_ng.sh New Generation ;-) shell script. As before, a diff with Bill's
      knowngood is provided when done. so now you can

        make test

      like everybody else out there!

    * Included VT and other fixes reported in the CRM114 ML these last few days:

         Hyperspace featurecounts Fixed

         TRE interface/regex cache fixed (Paolo) -- I discarded my older fix,
                                                    because I have not been able
                                                    to guarantee that it
                                                    would've worked, where
                                                    Paolo's does - has been
                                                    tested on a victim.

    * Updated and reformatted (indent cleanup) QUICKREF.txt.

    * QUICKREF.txt now includes preliminary documentation for the new CRM statements:

        cssanalyze <flags> (:report:) (:c1:) /params/
        cssbackup <flags> (:report:) (:dst: :c1:) /params/
        csscreate <flags> (:report:) (:c1:) /params/
        cssdiff <flags> (:report:) (:c1: :c2:) /params/
        cssinfo <flags> (:report:) (:c1:) /params/
        cssmerge <flags> (:report:) (destfile) (:c1:...:cN:) /params/
        cssrestore <flags> (:report:) (:c1: :src:) /params/

      These are intended to obsolete the external css* tools in due time and should
      be able to handle each classifier (see the bottom of classifier sources for the
      preliminary hooks/code).

+++ src/crm_debugger.c
@@ -443,8 +443,8 @@ long crm_debugger(void)
                 if (i == 0)
                 {
                     //    maybe the user put in a label?
-                    long tstart;
-                    long tlen;
+                    int tstart;
+                    int tlen;
                     crm_nextword(&inbuf[1], strlen(&inbuf[1]), 0,
                             &tstart, &tlen);
                     memmove(inbuf, &inbuf[1 + tstart], tlen);
@@ -490,8 +490,8 @@ long crm_debugger(void)
                 if (i == 0)
                 {
                     //    maybe the user put in a label?
-                    long tstart;
-                    long tlen;
+                    int tstart;
+                    int tlen;
                     crm_nextword(&inbuf[1], strlen(&inbuf[1]), 0,
                             &tstart, &tlen);
                     memmove(inbuf, &inbuf[1 + tstart], tlen);
@@ -536,9 +536,9 @@ long crm_debugger(void)
             {
                 //  do a debugger-level alteration
                 //    maybe the user put in a label?
-                long vstart, vlen;
-                long vindex;
-                long ostart, oend, olen;
+                int vstart, vlen;
+                int vindex;
+                int ostart, oend, olen;
                 crm_nextword(&inbuf[1], strlen(&inbuf[1]), 0,
                         &vstart, &vlen);
                 memmove(inbuf, &inbuf[1 + vstart], vlen);
